added semi-functional dropdown that allows user to select desired listings per page. I also added results
